requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain-core==0.3.72
langgraph-sdk==0.2.0
langsmith==0.4.8
llvmlite==0.44.0
matplotlib==3.10.1
numba==0.61.0
numpy==2.1.3
orjson==3.11.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
zstandard==0.23.0
